add test & source structure
