+++ .gitignore
@@ -1,1 +1,5 @@
-**/node_modules
+node_modules/
+build/
+
+*.zip
+!test/**/*.zip
