+++ package.json
@@ -1,12 +1,20 @@
 {
+  "name": "ts-zip",
+  "license": "UNLICENSED",
+  "scripts": {
+    "start": "nodemon src -x ts-node -w src",
+    "test": "jest"
+  },
   "dependencies": {},
   "devDependencies": {
+    "@types/chai": "^4.2.14",
+    "@types/jest": "^26.0.19",
     "@types/node": "^14.14.12",
+    "chai": "^4.2.0",
+    "jest": "^26.6.3",
     "nodemon": "^2.0.6",
+    "ts-jest": "^26.4.4",
     "ts-node": "^9.1.1",
     "typescript": "^4.1.3"
-  },
-  "scripts": {
-    "start": "nodemon index -e ts"
   }
 }
