+++ src/index.ts
@@ -1,0 +1,4 @@
+import { unzip } from "./unzip";
+import * as path from "path";
+
+unzip(path.join(__dirname, "../test/test.zip"));
